Realizada corrección en la bbdd, necesitamos realizar cálculos con el iva, lo pasamos de int a double

+++ WebAPI/Migrations/GPEContextModelSnapshot.cs
@@ -41,8 +41,8 @@ namespace GPE.Migrations
                     b.Property<bool>("Enabled")
                         .HasColumnType("tinyint(1)");
 
-                    b.Property<int>("Iva")
-                        .HasColumnType("int");
+                    b.Property<double>("Iva")
+                        .HasColumnType("double");
 
                     b.Property<double>("Price")
                         .HasColumnType("double");
@@ -59,7 +59,7 @@ namespace GPE.Migrations
                             Category = "RialOne",
                             Description = "PrimerArticuloToFlama",
                             Enabled = false,
-                            Iva = 21,
+                            Iva = 21.0,
                             Price = 10.5
                         },
                         new
@@ -69,7 +69,7 @@ namespace GPE.Migrations
                             Category = "RialOne",
                             Description = "SegundoArticuloToFlama",
                             Enabled = false,
-                            Iva = 4,
+                            Iva = 4.0,
                             Price = 15.5
                         });
                 });
@@ -351,8 +351,8 @@ namespace GPE.Migrations
                     b.Property<int>("Discount")
                         .HasColumnType("int");
 
-                    b.Property<int>("Iva")
-                        .HasColumnType("int");
+                    b.Property<double>("Iva")
+                        .HasColumnType("double");
 
                     b.Property<string>("LotId")
                         .IsRequired()
@@ -384,7 +384,7 @@ namespace GPE.Migrations
                             Category = "RialOne",
                             Description = "PrimerArticuloToFlama",
                             Discount = 0,
-                            Iva = 21,
+                            Iva = 21.0,
                             LotId = "Lot-01",
                             Price = 10.5,
                             Quantity = 15,
@@ -399,7 +399,7 @@ namespace GPE.Migrations
                             Category = "RialOne",
                             Description = "PrimerArticuloToFlama",
                             Discount = 10,
-                            Iva = 21,
+                            Iva = 21.0,
                             LotId = "Lot-01",
                             Price = 10.5,
                             Quantity = 25,
@@ -414,7 +414,7 @@ namespace GPE.Migrations
                             Category = "RialOne",
                             Description = "PrimerArticuloToFlama",
                             Discount = 0,
-                            Iva = 4,
+                            Iva = 4.0,
                             LotId = "Lot-02",
                             Price = 15.5,
                             Quantity = 15,
@@ -429,7 +429,7 @@ namespace GPE.Migrations
                             Category = "RialOne",
                             Description = "PrimerArticuloToFlama",
                             Discount = 10,
-                            Iva = 4,
+                            Iva = 4.0,
                             LotId = "Lot-02",
                             Price = 15.5,
                             Quantity = 25,
